+++ src/main/java/dat3/adventureXP/repository/EquipmentStatusRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Date;
 import java.util.Optional;
 
-public interface EquipmentStatusRepository extends JpaRepository<EquipmentStatus, Long> {
+public interface EquipmentStatusRepository extends JpaRepository<EquipmentStatus, Integer> {
     public Optional<EquipmentStatus> findByDate(Date date);
 }
